requirements.txt
# Core Dependencies
fastapi>=0.100
uvicorn[standard]>=0.23
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Frontend
streamlit>=1.28
plotly>=5.0

# LLM & AI
langchain>=0.0.300
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0

# API Integrations
arxiv>=1.4
requests>=2.31
PyGithub>=2.0
huggingface-hub>=0.16

# Utilities
python-dotenv>=1.0

# Testing
pytest>=7.0
pytest-cov>=4.0
httpx>=0.24
